Extract caller authentication into imbi-common
imbi-scheduler needs to authenticate bearer tokens and resolve
permissions from the graph (W6, PRD section 10). The PRD says that
happens "via imbi.common", but it did not: the implementation was ~800
lines in imbi.api.auth.permissions, coupled to imbi.api.models and
imbi.api.settings. imbi-assistant had already hit this and duplicated a
slim JWT-only copy, so the scheduler would have been the third.

Moved to imbi-common:

- The token path itself -- AuthContext, the JWT and ik_ API-key
  branches, the API-key LRU cache, permission loading, and
  require_permission -- as imbi.common.auth.permissions. The
  resource-level CAN_ACCESS helpers stay in imbi-api, where the
  resource_type -> AGE label map is meaningful.
- User, ServiceAccount, OrganizationEdge, MembershipProperties, and
  parse_scopes, since the shared path needs the principal types. This
  reverses their earlier move out of imbi-common, on the grounds that
  changed: three members authenticate callers now, not one.
- Argon2 hashing, because API-key secrets are verified with the same
  hasher.
- Auth.last_used_throttle_seconds, read by the shared API-key path.

imbi.api and imbi.assistant re-export every moved name, so the ~50
endpoint modules and 91 AuthContext construction sites are untouched and
resolve to the same class objects.

Two things needed care. Auth settings resolve through a registered
provider rather than a direct call: imbi-api extends Auth and keeps its
own singleton, and two singletons would each auto-generate a JWT secret
when IMBI_AUTH_JWT_SECRET is unset, so tokens minted against one would
fail against the other. The provider is a lambda, which also keeps the
~100 tests that patch imbi.api.settings.get_auth_settings effective.
Tests covering the moved code moved with it, so neither member's
coverage gate reports on code it no longer owns.

Verified: 1113 common, 2507 api, 195 assistant tests pass; coverage
91.27 / 85.99 / 95.64 percent; lint, format, basedpyright clean for all
three; mypy clean across 292 files. The only api test edits were import
paths and patch targets, as a behavior-preserving move should be.

Two behavior changes in imbi-assistant, both toward the platform norm:
it now accepts ik_ keys and service-account tokens, and role
inheritance resolves correctly (its local query walked INHERITS_FROM*0..,
which AGE does not honor, so only a role's own grants counted).
