+++ apps/assistant/src/imbi/assistant/auth.py
@@ -1,149 +1,36 @@
-"""Authentication and authorization for the assistant service."""
+"""Authentication and authorization for the assistant service.
 
-import logging
-import typing
+This module used to carry its own JWT verification and permission
+query, written before there was a shared one. Both now live in
+:mod:`imbi.common.auth.permissions`, so the names below are re-exports
+and the endpoints that import them are unchanged.
 
-import fastapi
-import jwt
-import pydantic
-from fastapi import security
+Two behaviors changed with the switch, both toward the platform norm:
 
-from imbi.common import graph, settings
-from imbi.common.auth import core
+* A caller may now authenticate with an ``ik_`` API key or a
+  service-account client-credentials token, not only a user JWT.
+  ``AuthContext.user`` is ``None`` for a service account, which is why
+  the endpoints here reach for ``require_user`` rather than ``user``.
+* Role inheritance resolves correctly. The local query walked
+  ``INHERITS_FROM*0..``, and Apache AGE does not honor the zero-hop
+  case, so a role's own grants were the only ones that counted. The
+  shared query collects ancestors with ``*1..`` and unions the start
+  role back in.
+"""
 
-LOGGER = logging.getLogger(__name__)
+from imbi.common import models
+from imbi.common.auth import permissions
 
-oauth2_scheme = security.HTTPBearer(auto_error=False)
+__all__ = [
+    'AuthContext',
+    'User',
+    'get_current_user',
+    'load_principal_permissions',
+    'oauth2_scheme',
+]
 
-
-class User(pydantic.BaseModel):
-    """Minimal user model for assistant endpoints."""
-
-    model_config = pydantic.ConfigDict(extra='ignore')
-
-    email: pydantic.EmailStr
-    display_name: str
-    is_active: bool = True
-    is_admin: bool = False
-
-
-class AuthContext(pydantic.BaseModel):
-    """Authentication context for the current request."""
-
-    user: User | None = None
-    session_id: str | None = None
-    auth_method: typing.Literal['jwt'] = 'jwt'
-    permissions: set[str] = pydantic.Field(
-        default_factory=set,
-    )
-
-    @property
-    def is_admin(self) -> bool:
-        """Return whether the principal is an admin."""
-        return self.user.is_admin if self.user else False
-
-    @property
-    def require_user(self) -> User:
-        """Return the authenticated user or raise 403."""
-        if self.user is None:
-            raise fastapi.HTTPException(
-                403,
-                'This endpoint requires user authentication',
-            )
-        return self.user
-
-
-async def load_user_permissions(
-    db: graph.Graph,
-    email: str,
-) -> set[str]:
-    """Get permission names granted to a user."""
-    query = """
-    MATCH (u:User {{email: {email}}})
-          -[m:MEMBER_OF]->(o:Organization)
-    MATCH (r:Role {{slug: m.role}})
-    OPTIONAL MATCH (r)-[:INHERITS_FROM*0..]->(parent:Role)
-    WITH DISTINCT parent
-    OPTIONAL MATCH (parent)-[:GRANTS]->(perm:Permission)
-    RETURN collect(DISTINCT perm.name) AS permissions
-    """
-    records = await db.execute(
-        query,
-        {'email': email},
-        ['permissions'],
-    )
-    if not records:
-        return set()
-    perms = graph.parse_agtype(records[0].get('permissions'))
-    if isinstance(perms, list):
-        return {str(p) for p in typing.cast('list[object]', perms)}
-    return set()
-
-
-async def get_current_user(
-    db: graph.Pool,
-    credentials: (
-        security.HTTPAuthorizationCredentials | None
-    ) = fastapi.Depends(oauth2_scheme),  # noqa: B008
-) -> AuthContext:
-    """FastAPI dependency to get the current authenticated
-    user."""
-    if not credentials:
-        raise fastapi.HTTPException(
-            status_code=401,
-            detail='Missing authentication credentials',
-            headers={'WWW-Authenticate': 'Bearer'},
-        )
-
-    auth_settings = settings.get_auth_settings()
-    token = credentials.credentials
-
-    try:
-        claims = core.verify_token(token, auth_settings)
-    except jwt.ExpiredSignatureError as err:
-        raise fastapi.HTTPException(
-            status_code=401, detail='Token has expired'
-        ) from err
-    except jwt.InvalidTokenError as err:
-        raise fastapi.HTTPException(
-            status_code=401, detail='Invalid token'
-        ) from err
-
-    if claims.get('type') != 'access':
-        raise fastapi.HTTPException(
-            status_code=401, detail='Invalid token type'
-        )
-
-    subject = claims.get('sub')
-    if not subject:
-        raise fastapi.HTTPException(
-            status_code=401, detail='Token missing subject'
-        )
-
-    user_query = """
-    MATCH (u:User {{email: {email}}})
-    RETURN u
-    """
-    records = await db.execute(
-        user_query,
-        {'email': subject},
-        ['u'],
-    )
-    if not records:
-        raise fastapi.HTTPException(status_code=401, detail='User not found')
-    user_data = graph.parse_agtype(records[0]['u'])
-    user = User(**user_data)
-
-    if not user.is_active:
-        raise fastapi.HTTPException(
-            status_code=401,
-            detail='User account is inactive',
-        )
-
-    perms = await load_user_permissions(db, subject)
-
-    return AuthContext(
-        user=user,
-        session_id=claims.get('jti'),
-        permissions=perms,
-    )
+AuthContext = permissions.AuthContext
+User = models.User
+get_current_user = permissions.get_current_user
+load_principal_permissions = permissions.load_principal_permissions
+oauth2_scheme = permissions.oauth2_scheme
