+++ apps/assistant/tests/test_auth.py
@@ -1,274 +1,76 @@
-"""Tests for assistant auth module."""
+"""Tests for the assistant's auth re-exports.
+
+The JWT verification, principal loading, and permission query this
+module used to own now live in :mod:`imbi.common.auth.permissions`, and
+their cases live in ``libraries/common/tests/auth/``. What is worth
+testing here is the seam: that the names the endpoints import still
+resolve, and that they resolve to the shared implementation rather than
+to a copy that could drift.
+"""
 
 import unittest
-from unittest import mock
 
 import fastapi
 
 from imbi.assistant import auth
+from imbi.common import models
+from imbi.common.auth import permissions
 
 
-class UserModelTestCase(unittest.TestCase):
-    """Test cases for User model."""
+class ReExportTestCase(unittest.TestCase):
+    """The endpoints import these names from ``assistant.auth``."""
 
-    def test_create_user(self) -> None:
-        user = auth.User(
-            email='test@example.com',
-            display_name='Test User',
+    def test_auth_context_is_the_shared_one(self) -> None:
+        self.assertIs(auth.AuthContext, permissions.AuthContext)
+
+    def test_user_is_the_shared_model(self) -> None:
+        self.assertIs(auth.User, models.User)
+
+    def test_get_current_user_is_the_shared_dependency(self) -> None:
+        self.assertIs(auth.get_current_user, permissions.get_current_user)
+
+    def test_oauth2_scheme_is_the_shared_scheme(self) -> None:
+        self.assertIs(auth.oauth2_scheme, permissions.oauth2_scheme)
+
+    def test_permission_loader_is_the_shared_query(self) -> None:
+        self.assertIs(
+            auth.load_principal_permissions,
+            permissions.load_principal_permissions,
         )
-        self.assertEqual(user.email, 'test@example.com')
-        self.assertEqual(user.display_name, 'Test User')
-        self.assertTrue(user.is_active)
-        self.assertFalse(user.is_admin)
 
-    def test_create_admin_user(self) -> None:
-        user = auth.User(
-            email='admin@example.com',
-            display_name='Admin',
-            is_admin=True,
-        )
-        self.assertTrue(user.is_admin)
+    def test_every_exported_name_exists(self) -> None:
+        for name in auth.__all__:
+            with self.subTest(name=name):
+                self.assertTrue(hasattr(auth, name))
 
 
-class AuthContextTestCase(unittest.TestCase):
-    """Test cases for AuthContext."""
+class ServiceAccountContextTestCase(unittest.TestCase):
+    """A service-account caller has no ``user``.
 
-    def test_is_admin_with_admin_user(self) -> None:
-        user = auth.User(
-            email='admin@example.com',
-            display_name='Admin',
-            is_admin=True,
-        )
-        ctx = auth.AuthContext(user=user)
-        self.assertTrue(ctx.is_admin)
+    Accepting client-credentials tokens is new here -- the local
+    implementation only ever built a context from a user JWT. The
+    endpoints reach for ``require_user``, which raises 403 rather than
+    returning None, so this is the behavior that keeps a service-account
+    token out of code that assumes a person.
+    """
 
-    def test_is_admin_without_user(self) -> None:
-        ctx = auth.AuthContext()
-        self.assertFalse(ctx.is_admin)
-
-    def test_require_user_returns_user(self) -> None:
-        user = auth.User(
-            email='test@example.com',
-            display_name='Test',
-        )
-        ctx = auth.AuthContext(user=user)
-        self.assertIs(ctx.require_user, user)
-
-    def test_require_user_raises_without_user(
-        self,
-    ) -> None:
-        ctx = auth.AuthContext()
-        with self.assertRaises(fastapi.HTTPException) as exc:
-            ctx.require_user
-        self.assertEqual(exc.exception.status_code, 403)
-
-    def test_permissions_default_empty(self) -> None:
-        ctx = auth.AuthContext()
-        self.assertEqual(ctx.permissions, set())
-
-
-class LoadUserPermissionsTestCase(
-    unittest.IsolatedAsyncioTestCase,
-):
-    """Test cases for load_user_permissions."""
-
-    async def test_load_permissions(self) -> None:
-        db = mock.AsyncMock()
-        db.execute.return_value = [
-            {
-                'permissions': [
-                    'project:read',
-                    'team:read',
-                ]
-            }
-        ]
-        with mock.patch(
-            'imbi.common.graph.parse_agtype',
-            return_value=[
-                'project:read',
-                'team:read',
-            ],
-        ):
-            perms = await auth.load_user_permissions(db, 'test@example.com')
-        self.assertEqual(perms, {'project:read', 'team:read'})
-
-    async def test_load_permissions_empty(self) -> None:
-        db = mock.AsyncMock()
-        db.execute.return_value = []
-        perms = await auth.load_user_permissions(db, 'test@example.com')
-        self.assertEqual(perms, set())
-
-
-class GetCurrentUserTestCase(
-    unittest.IsolatedAsyncioTestCase,
-):
-    """Test cases for get_current_user."""
-
-    async def test_missing_credentials(self) -> None:
-        db = mock.AsyncMock()
-        with self.assertRaises(
-            fastapi.HTTPException,
-        ) as ctx:
-            await auth.get_current_user(db, None)
-        self.assertEqual(ctx.exception.status_code, 401)
-
-    async def test_expired_token(self) -> None:
-        import jwt
-
-        db = mock.AsyncMock()
-        creds = mock.MagicMock()
-        creds.credentials = 'expired-token'
-        with mock.patch(
-            'imbi.common.auth.core.verify_token',
-            side_effect=jwt.ExpiredSignatureError(),
-        ):
-            with self.assertRaises(
-                fastapi.HTTPException,
-            ) as ctx:
-                await auth.get_current_user(db, creds)
-            self.assertEqual(ctx.exception.status_code, 401)
-            self.assertIn('expired', str(ctx.exception.detail))
-
-    async def test_invalid_token(self) -> None:
-        import jwt
-
-        db = mock.AsyncMock()
-        creds = mock.MagicMock()
-        creds.credentials = 'bad-token'
-        with mock.patch(
-            'imbi.common.auth.core.verify_token',
-            side_effect=jwt.InvalidTokenError(),
-        ):
-            with self.assertRaises(
-                fastapi.HTTPException,
-            ) as ctx:
-                await auth.get_current_user(db, creds)
-            self.assertEqual(ctx.exception.status_code, 401)
-
-    async def test_invalid_token_type(self) -> None:
-        db = mock.AsyncMock()
-        creds = mock.MagicMock()
-        creds.credentials = 'some-token'
-        with mock.patch(
-            'imbi.common.auth.core.verify_token',
-            return_value={
-                'type': 'refresh',
-                'sub': 'x',
-            },
-        ):
-            with self.assertRaises(
-                fastapi.HTTPException,
-            ) as ctx:
-                await auth.get_current_user(db, creds)
-            self.assertEqual(ctx.exception.status_code, 401)
-            self.assertIn('token type', str(ctx.exception.detail))
-
-    async def test_missing_subject(self) -> None:
-        db = mock.AsyncMock()
-        creds = mock.MagicMock()
-        creds.credentials = 'some-token'
-        with mock.patch(
-            'imbi.common.auth.core.verify_token',
-            return_value={'type': 'access'},
-        ):
-            with self.assertRaises(
-                fastapi.HTTPException,
-            ) as ctx:
-                await auth.get_current_user(db, creds)
-            self.assertEqual(ctx.exception.status_code, 401)
-            self.assertIn('subject', str(ctx.exception.detail))
-
-    async def test_user_not_found(self) -> None:
-        db = mock.AsyncMock()
-        db.execute.return_value = []
-        creds = mock.MagicMock()
-        creds.credentials = 'some-token'
-        with mock.patch(
-            'imbi.common.auth.core.verify_token',
-            return_value={
-                'type': 'access',
-                'sub': 'nobody@example.com',
-                'jti': 'j1',
-            },
-        ):
-            with self.assertRaises(
-                fastapi.HTTPException,
-            ) as ctx:
-                await auth.get_current_user(db, creds)
-            self.assertEqual(ctx.exception.status_code, 401)
-
-    async def test_inactive_user(self) -> None:
-        db = mock.AsyncMock()
-        creds = mock.MagicMock()
-        creds.credentials = 'some-token'
-        user_data = {
-            'email': 'inactive@example.com',
-            'display_name': 'Inactive',
-            'is_active': False,
-            'is_admin': False,
-        }
-        db.execute.return_value = [{'u': user_data}]
-        with (
-            mock.patch(
-                'imbi.common.auth.core.verify_token',
-                return_value={
-                    'type': 'access',
-                    'sub': 'inactive@example.com',
-                    'jti': 'j1',
-                },
+    def _context(self) -> auth.AuthContext:
+        return auth.AuthContext(
+            service_account=models.ServiceAccount(
+                slug='imbi-scheduler', display_name='Scheduler'
             ),
-            mock.patch(
-                'imbi.common.graph.parse_agtype',
-                side_effect=lambda x: x,
-            ),
-        ):
-            with self.assertRaises(
-                fastapi.HTTPException,
-            ) as ctx:
-                await auth.get_current_user(db, creds)
-            self.assertEqual(ctx.exception.status_code, 401)
-            self.assertIn('inactive', str(ctx.exception.detail))
-
-    async def test_successful_auth(self) -> None:
-        db = mock.AsyncMock()
-        creds = mock.MagicMock()
-        creds.credentials = 'good-token'
-
-        user_data = {
-            'email': 'test@example.com',
-            'display_name': 'Test User',
-            'is_active': True,
-            'is_admin': False,
-        }
-
-        # First call: user lookup; second: permissions
-        db.execute.side_effect = [
-            [{'u': user_data}],
-            [{'permissions': ['project:read']}],
-        ]
-
-        with (
-            mock.patch(
-                'imbi.common.auth.core.verify_token',
-                return_value={
-                    'type': 'access',
-                    'sub': 'test@example.com',
-                    'jti': 'j1',
-                },
-            ),
-            mock.patch(
-                'imbi.common.graph.parse_agtype',
-                side_effect=[
-                    user_data,
-                    ['project:read'],
-                ],
-            ),
-        ):
-            result = await auth.get_current_user(db, creds)
-        self.assertIsInstance(result, auth.AuthContext)
-        self.assertEqual(
-            result.require_user.email,
-            'test@example.com',
+            auth_method='client_credentials',
         )
-        self.assertEqual(result.permissions, {'project:read'})
-        self.assertEqual(result.session_id, 'j1')
+
+    def test_require_user_rejects_a_service_account(self) -> None:
+        ctx = self._context()
+        self.assertIsNone(ctx.user)
+        with self.assertRaises(fastapi.HTTPException) as caught:
+            _ = ctx.require_user
+        self.assertEqual(caught.exception.status_code, 403)
+
+    def test_principal_name_falls_back_to_the_slug(self) -> None:
+        self.assertEqual(self._context().principal_name, 'imbi-scheduler')
+
+    def test_a_service_account_is_not_an_admin(self) -> None:
+        self.assertFalse(self._context().is_admin)
