+++ libraries/common/pyproject.toml
@@ -18,8 +18,10 @@ classifiers = [
   "Programming Language :: Python :: 3.14",
 ]
 dependencies = [
+  "argon2-cffi>=23.1",
   "cryptography>=42.0.4",
   "httpx>=0.28.1,<1",
+  "jsonpatch>=1.33",
   "jsonpointer>=3.1.1",
   "jsonschema-models",
   "markdown-it-py>=3.0",
